Remove debug code, add comments
