+++ packages/walt-compiler/src/semantics/map-function/index.js
@@ -9,7 +9,6 @@
 import Syntax from "../../Syntax";
 import curry from "curry";
 import mapNode from "../../utils/map-node";
-import { stringEncoder } from "../../utils/string";
 import { parseDeclaration } from "./declaration";
 import mapCharacterLiteral from "../map-char";
 import mapUnary from "../map-unary";
@@ -27,14 +26,7 @@ import makePair from "./map-pair";
 import walkNode from "../../utils/walk-node";
 import { balanceTypesInMathExpression } from "./patch-typecasts";
 import { collapseClosureIdentifier, CLOSURE_BASE } from "../closure";
-import { generateStringOffset } from "../../utils/string";
-import {
-  FUNCTION_INDEX,
-  CLOSURE_TYPE,
-  FUNCTION_METADATA,
-  STATIC_STRING,
-  STATIC_TOTAL_LENGTH,
-} from "../metadata";
+import { FUNCTION_INDEX, CLOSURE_TYPE, FUNCTION_METADATA } from "../metadata";
 import type { NodeType } from "../../flow/types";
 
 /**
@@ -211,34 +203,19 @@ const mapFunctionNode = (options, node, topLevelTransform) => {
       });
     },
     [Syntax.CharacterLiteral]: mapCharacterLiteral,
-    [Syntax.StringLiteral]: (stringLiteral, transform) => {
+    [Syntax.StringLiteral]: (stringLiteral, _) => {
       const { statics } = options;
       const { value } = stringLiteral;
 
       // did we already encode the static?
       if (!(value in statics)) {
-        const offset = statics[STATIC_TOTAL_LENGTH];
-        statics[STATIC_TOTAL_LENGTH] += value.length;
-        statics[value] = offset;
-        console.log(value, value.length, offset);
+        statics[value] = null;
       }
 
+      // It's too early to tranform a string at this point
+      // we need additional information, only available in the generator.
+      // This also avoids doing the work in two places, in semantics AND gen
       return stringLiteral;
-      // const transformed = transform({
-      //   ...stringLiteral,
-      //   value: String(
-      //     generateStringOffset(statics, value, STATIC_TOTAL_LENGTH)
-      //   ),
-      //   meta: {
-      //     ...stringLiteral.meta,
-      //     [STATIC_STRING]: stringEncoder(value),
-      //   },
-      //   Type: Syntax.Constant,
-      // });
-      // // it really does not matter when we write the key in
-      // statics[stringLiteral.value] = transformed;
-
-      // return transformed;
     },
     [Syntax.Assignment]: mapAssignment,
     [Syntax.MemoryAssignment]: (inputNode, transform) => {
