+++ packages/walt-compiler/src/semantics/metadata.js
@@ -15,8 +15,10 @@ export const OBJECT_SIZE = "object/size";
 export const TYPE_CAST = "type/cast";
 export const OBJECT_KEY_TYPES = "object/key-types";
 export const CLOSURE_TYPE = "closure/type";
-export const AST_METADATA = "@@global/ast";
+export const AST_METADATA = "AST_METADATA";
 export const FUNCTION_METADATA = "@@function/meta";
 export const ALIAS = "alias";
+
+// Statics
 export const STATIC_STRING = "static/string";
 export const STATIC_INDEX = "static/index";
